+++ realm/realm-library/src/main/cpp/io_realm_SyncSession.cpp
@@ -23,7 +23,6 @@
 
 #include "util.hpp"
 
-using namespace std;
 using namespace realm;
 using namespace sync;
 
@@ -48,4 +47,42 @@ JNIEXPORT jboolean JNICALL Java_io_realm_SyncSession_nativeRefreshAccessToken(JN
     }
     CATCH_STD()
     return JNI_FALSE;
+}
+
+
+JNIEXPORT jlong JNICALL Java_io_realm_SyncSession_nativeAddProgressListener(JNIEnv* env, jclass,
+                                                                            jstring localRealmPath,
+                                                                            jobject listenerWrapper, jint direction,
+                                                                            jboolean isStreaming)
+{
+    try {
+        JStringAccessor local_realm_path(env, localRealmPath);
+        std::shared_ptr<SyncSession> session = SyncManager::shared().get_existing_active_session(local_realm_path);
+        if (!session) {
+            ThrowException(env, IllegalState, "FIXME: Cannot register a progress listener before a session is "
+                                              "created. Only happens during Realm.getInstance() right now.");
+            return static_cast<jlong>(0);
+        }
+
+        SyncSession::NotifierType type =
+            direction == (jint == 1) ? SyncSession::NotifierType::download : SyncSession::NotifierType::upload;
+
+        std::function<SyncProgressNotifierCallback> callback = [=](uint64_t transferred, uint64_t transferrable) {
+            // Callback to Java
+        }
+        uint64_t token = session->register_progress_notifier(callback, type, to_bool(isStreaming));
+        return static_cast<jlong>(token);
+    }
+    CATCH_STD()
+}
+
+JNIEXPORT void JNICALL Java_io_realm_SyncSession_nativeRemoveProgressListener(JNIEnv* env, jclass,
+                                                                              jstring localRealmPath,
+                                                                              jlong listenerToken)
+{
+    JStringAccessor local_realm_path(env, localRealmPath);
+    std::shared_ptr<SyncSession> session = SyncManager::shared().get_existing_active_session(local_realm_path);
+    if (session) {
+        session->unregister_progress_notifier(static_cast<uint64_t>(listenerToken));
+    }
 }